Separate React shared
