+++ packages/react-cosmos-shared/src/index.js
@@ -1,4 +1,3 @@
 export { default as importModule } from './import-module';
 export { default as createLinkedList } from './linked-list';
-export { default as proxyPropTypes } from './proxy-prop-types';
 export { default as splitUnserializableParts } from './unserializable-parts';
