+++ packages/react-cosmos-shared/src/react/index.js
@@ -1,0 +1,1 @@
+export { default as proxyPropTypes } from './proxy-prop-types';
