+++ packages/react-cosmos-xhr-proxy/src/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import xhrMock from 'xhr-mock';
-import { proxyPropTypes } from 'react-cosmos-shared';
+import { proxyPropTypes } from 'react-cosmos-shared/lib/react';
 
 const defaults = {
   fixtureKey: 'xhr'
